RF: module rename
