+++ permute/stratified.py
